project/requirements.txt
aniso8601==9.0.1
asgiref==3.5.2
backports.zoneinfo==0.2.1
cffi==1.15.1
cryptography==39.0.0
Django==4.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
graphene==3.2.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
promise==2.3
pycparser==2.21
PyJWT==2.6.0
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
